[RESTEASY-1721] Eliminate ContextParameterInjector's dependence on servlet classes
[RESTEASY-1721] Remove another reference to ServletOutputStream.
