+++ resteasy-core/src/main/java/org/jboss/resteasy/core/ContextParameterInjector.java
@@ -9,15 +9,13 @@ import org.jboss.resteasy.spi.ResteasyProviderFactory;
 import org.jboss.resteasy.spi.ValueInjector;
 import org.jboss.resteasy.spi.util.Types;
 
-import javax.servlet.ServletOutputStream;
-import javax.servlet.WriteListener;
 import javax.ws.rs.container.ResourceInfo;
 import javax.ws.rs.core.Application;
 import javax.ws.rs.ext.Providers;
 import javax.ws.rs.sse.Sse;
 import javax.ws.rs.sse.SseEventSink;
 
-import java.io.IOException;
+import java.io.OutputStream;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.InvocationTargetException;
@@ -38,12 +36,37 @@ import java.util.concurrent.CompletionStage;
 @SuppressWarnings("unchecked")
 public class ContextParameterInjector implements ValueInjector
 {
+   private static boolean servletPresent;
+
    private Class<?> rawType;
    private Class<?> proxy;
    private ResteasyProviderFactory factory;
    private Type genericType;
    private Annotation[] annotations;
    private volatile boolean outputStreamWasWritten = false;
+
+   static
+   {
+      try
+      {
+         Class.forName("javax.servlet.http.HttpServletResponse");
+         servletPresent = true;
+      }
+      catch (Throwable e)
+      {
+        //
+      }
+   }
+
+   public boolean isOutputStreamWasWritten()
+   {
+      return outputStreamWasWritten;
+   }
+
+   public void setOutputStreamWasWritten(boolean outputStreamWasWritten)
+   {
+      this.outputStreamWasWritten = outputStreamWasWritten;
+   }
 
    public ContextParameterInjector(final Class<?> proxy, final Class<?> rawType, final Type genericType, final Annotation[] annotations, final ResteasyProviderFactory factory)
    {
@@ -135,8 +158,8 @@ public class ContextParameterInjector implements ValueInjector
             {
                if ("getOutputStream".equals(method.getName()))
                {
-                  ServletOutputStream sos = (ServletOutputStream) method.invoke(delegate, objects);
-                  return new ContextOutputStream(sos);
+                  OutputStream sos = (OutputStream) method.invoke(delegate, objects);
+                  return wrapServletOutputStream(sos);
                }
             }
             return method.invoke(delegate, objects);
@@ -241,149 +264,8 @@ public class ContextParameterInjector implements ValueInjector
       return set.toArray(new Class<?>[]{});
    }
 
-   private final class ContextOutputStream extends ServletOutputStream
+   OutputStream wrapServletOutputStream(OutputStream os)
    {
-      private ServletOutputStream delegate;
-
-      ContextOutputStream(final ServletOutputStream delegate)
-      {
-         this.delegate = delegate;
-      }
-
-      ////////////////////////////////////////////////////////////////
-      /// ServletOutputStream methods
-      ////////////////////////////////////////////////////////////////
-      @Override
-      public void print(String s) throws IOException
-      {
-         delegate.print(s);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void print(boolean b) throws IOException
-      {
-         delegate.print(b);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void print(char c) throws IOException
-      {
-         delegate.print(c);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void print(int i) throws IOException
-      {
-         delegate.print(i);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void print(long l) throws IOException
-      {
-         delegate.print(l);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void print(float f) throws IOException
-      {
-         delegate.print(f);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void print(double d) throws IOException
-      {
-         delegate.print(d);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void println() throws IOException
-      {
-         delegate.println();
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void println(String s) throws IOException
-      {
-         delegate.println(s);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void println(boolean b) throws IOException
-      {
-         delegate.println(b);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void println(char c) throws IOException
-      {
-         delegate.print(c);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void println(int i) throws IOException
-      {
-         delegate.println(i);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void println(long l) throws IOException
-      {
-         delegate.println(l);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void println(float f) throws IOException
-      {
-         delegate.println(f);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void println(double d) throws IOException
-      {
-         delegate.println(d);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public boolean isReady()
-      {
-         return delegate.isReady();
-      }
-      @Override
-      public void setWriteListener(WriteListener writeListener)
-      {
-         delegate.setWriteListener(writeListener);
-      }
-
-      ////////////////////////////////////////////////////////////////
-      /// OutputStream methods
-      ////////////////////////////////////////////////////////////////
-      @Override
-      public void write(byte[] b) throws IOException
-      {
-          delegate.write(b);
-          outputStreamWasWritten = true;
-      }
-      @Override
-      public void write(byte[] b, int off, int len) throws IOException
-      {
-         delegate.write(b, off, len);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void write(int b) throws IOException
-      {
-         delegate.write(b);
-         outputStreamWasWritten = true;
-      }
-      @Override
-      public void flush() throws IOException
-      {
-         delegate.flush();
-      }
-      @Override
-      public void close() throws IOException
-      {
-         delegate.close();
-      }
+      return servletPresent ? new ContextServletOutputStream(this, os) : os;
    }
 }
